Redo to follow better OOP
