+++ src/main/java/duke/task/Deadline.java
@@ -1,0 +1,15 @@
+package duke.task;
+
+public class Deadline extends Task {
+    protected String deadline;
+
+    public Deadline(String taskName, String deadline) {
+        super(taskName);
+        this.deadline = deadline;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + deadline + ")";
+    }
+}
