+++ src/main/java/duke/task/Event.java
@@ -1,0 +1,15 @@
+package duke.task;
+
+public class Event extends Task {
+    protected String event;
+
+    public Event(String taskName, String event) {
+        super(taskName);
+        this.event = event;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (at: " + event + ")";
+    }
+}
